plugins/gce_scanner/requirements.txt
# plugins/gce_scanner/requirements.txt

# Core dependencies
pika==1.3.2
requests==2.31.0
python-gvm>=24.8.0  # Latest version for GMP 22.x support

# XML processing
lxml==5.1.0
xmltodict==0.13.0

# Utilities
python-dateutil==2.8.2